Creates and implements basic AI structure for OnePartStatue
The OnePartStatue will not move until it spots a player. Then it will enter an attacking phase for a set duration. It will rotate to face the player and will hop attack until the duration is over and rests before it repeats. The attack duration, cooldown, rest duration, jump angle and impulse can be adjusted inside of the BT_OnePartStatue. The sight radius will have to be adjusted in the AIC_OnePartStatue.